Cleanup, window resizing and docs
