+++ crates/yakui-vulkan/Cargo.toml
@@ -9,13 +9,12 @@ version = "0.2.0"
 [dependencies]
 ash = {version = "0.37.2", features = ["linked"]}
 bytemuck = "1.13.0"
-glam = "0.22.0"
-raw-window-handle = "0.5.0"
 thunderdome = "0.6.0"
 yakui = {path = "../yakui"}
-yakui-core = {path = "../yakui-core", version = "0.2.0"}
+yakui-core = {path = "../yakui-core"}
 
 [dev-dependencies]
 ash-window = "0.12.0"
 image = "0.24.5"
+raw-window-handle = "0.5.0"
 winit = "0.27.5"
